serve/requirements.txt
fastapi
uvicorn[standard]
pillow
torch
torchvision
timm
python-multipart
openai
python-dotenv
pydantic-settings
transformers